feat: postcss-syntax should silence parse errors on demand (#641)
* feat: postcss-syntax should silence parse errors on demand

Since the postcss-syntax is primarily used for linting with Stylelint, there are two
areas where it can be used

1. CI pipeline check
2. IDE integration

For 1. we want to always throw but for 2. where users can be in the
process of editing their code, it's not always parseable so we want to
be able to silence here.

* Change files